requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-groq
python-dotenv
plotly
matplotlib
sqlalchemy